Split the codebase bit better. Make clear distinction between vals parser and the vm
