+++ src/collections/mod.rs
@@ -1,3 +1,1 @@
-mod stack;
-
-pub use stack::Stack;
+pub mod stack;
