--- src/iterators/mod.rs
@@ -1,5 +1,0 @@
-mod tokenizer;
-mod parser;
-
-pub use tokenizer::Tokenizer;
-pub use parser::Parser;
