+++ src/vm/sys_api.rs
@@ -1,0 +1,6 @@
+
+use std::collections::HashMap;
+use super::VM;
+
+pub type SysFunc = fn(vm: &mut VM);
+pub type SysApi = HashMap<String, SysFunc>;
